requirements.txt
accelerate==1.11.0
datasets==3.1.0
numpy>=1.24,<3
peft==0.18.0
torch==2.7.1
transformers==4.57.6
trl==0.24.0
bitsandbytes>=0.43.0
fastapi>=0.110.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
requests>=2.31.0
Pillow>=10.0.0
